LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-029-expected.txt

This text should be visible.

FAIL innerText on locked element. assert_equals: expected "This text should be visible." but got "This text should be visible.\n"
